requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
twilio==9.0.4
python-decouple==3.8